[UPDATED] Final 1

+++ Library-Web-App/Library-Web-App/Service/BookService.cs
@@ -3,129 +3,46 @@ using Library_Web_App.Data.Entities;
 using Library_Web_App.Data.ViewModels.Book;
 using Library_Web_App.Data.ViewModels.Comment;
 using Library_Web_App.Migrations;
+using Library_Web_App.Service.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using System.Linq;
 
 namespace Library_Web_App.Service
 {
-    public class BookService
+    public class BookService : IBookService
     {
         private readonly ApplicationContext context;
+        private readonly ILikeService likeService;
+        private readonly ICommentService commentService;
 
-        public BookService(ApplicationContext context)
+        public BookService(ApplicationContext context, ILikeService likeService, ICommentService commentService)
         {
             this.context = context;
+            this.likeService = likeService;
+            this.commentService = commentService;
         }
 
         public List<BookIndexViewModel> GetAll()
             => context.Books.Select(book => new BookIndexViewModel(book)).ToList();
-
-        public List<Book> GetAllByGenre(string genre)
-            => context.Books
-                      .Where(book => book.Genre == genre)
-                      .ToList();
-
-        public List<Book> GetAllByAuthor(string author)
-            => context.Books
-                      .Where(book => book.Author == author)
-                      .ToList();
-
-        public List<Book> GetAllByYearOfPublication(int yearOfPublication)
-            => context.Books
-                      .Where(book => book.YearOfPublication == yearOfPublication)
-                      .ToList();
 
         public Book GetById(int id)
             => context.Books
                       .FirstOrDefault(book => id == book.Id);
 
         public BookInfoViewModel GetByIdExtendedInfo(string userName, int id)
-            => new BookInfoViewModel(GetById(id), GetLikesCount(id), GetCommentsExtendedInfo(id), IsBookLikedByCurUser(id, userName));
-
-        public bool IsBookLikedByCurUser(int id, string userName)
-            => GetBookLikeMadeByUser(id, userName) != null;
-
-        public Like GetBookLikeMadeByUser(int bookId, string userName)
-            => context.Likes
-                      .Include(c => c.User)
-                      .FirstOrDefault(like => like.BookId == bookId && like.User.UserName == userName);
-
-        public List<Like> GetBookLikes(int id)
-            => context.Likes
-                      .Include(c => c.User)
-                      .Where(like => like.BookId == id)
-                      .ToList();
-
-        public int GetLikesCount(int id)
-            => GetBookLikes(id).Count();
-
-        public List<Comment> GetComments(int id)
-            => context.Comments
-                      .Include(c => c.User)
-                      .Where(comment => comment.BookId == id)
-                      .ToList();
-
-        public List<CommentInfoViewModel> GetCommentsExtendedInfo(int id)
-            => GetComments(id).Select(comment => new CommentInfoViewModel(comment, GetUserRoleColor(comment.UserId))).ToList();
-
-        public string GetUserRoleColor(string userId)
         {
-            string roleId = context.UserRoles
-                                   .FirstOrDefault(user_role => user_role.UserId == userId)
-                                   .RoleId;
+            Book book = GetById(id);
 
-            Role role = context.Roles
-                              .FirstOrDefault(role => role.Id == roleId);
+            // ако книга с подаденото id не е намерено
+            if (book == null)
+                return null;
 
-            return role.Color;
-        }
-
-        public Comment GetComment(int id)
-            => context.Comments
-                      .Include(c => c.User)
-                      .FirstOrDefault(comment => comment.Id == id);
-
-        public void AddComment(int bookId, string userName, string message)
-        {
-            context.Comments.Add(new Comment(bookId, GetUser(userName).Id, message, DateTime.Now));
-            context.SaveChanges();
-        }
-
-        public void DeleteAllBookLikes(int id)
-        {
-            List<Like> likes = GetBookLikes(id);
+            int likeCount = likeService.GetLikesCount(id);
+            List<CommentInfoViewModel> comments = commentService.GetCommentsExtendedInfo(id);
+            bool isLiked = likeService.IsBookLikedByCurUser(id, userName);
             
-            foreach(Like like in likes)
-                context.Likes.Remove(like);
-
-            context.SaveChanges();
+            return new BookInfoViewModel(book, likeCount, comments, isLiked);
         }
-
-        public void DeleteAllBookComments(int id)
-        {
-            List<Comment> comments = GetComments(id);
-
-            foreach (Comment comment in comments)
-                context.Comments.Remove(comment);
-
-            context.SaveChanges();
-        }
-
-        public int DeleteCommentById(int id)
-        {
-            var comment = GetComment(id);
-            int bookId = comment.BookId;
-
-            context.Comments.Remove(comment);
-            context.SaveChanges();
-
-            return bookId;
-        }
-
-        public User GetUser(string userName)
-            => context.Users
-                      .FirstOrDefault(user => user.UserName == userName);
-
         public void Add(Book book)
         {
             context.Books.Add(book);
@@ -142,23 +59,27 @@ namespace Library_Web_App.Service
         {
             Book book = GetById(id);
 
+            if (book == null)
+                return;
+
             context.Books.Remove(book);
-            DeleteAllBookLikes(id);
-            DeleteAllBookComments(id);
+
+            likeService.DeleteAllBookLikes(id);
+            commentService.DeleteAllBookComments(id);
             
             context.SaveChanges();
         }
 
         public void Like(int bookId, string userName)
-        {
-            context.Likes.Add(new Like(bookId, GetUser(userName).Id));
-            context.SaveChanges();
-        }
+            => likeService.Like(bookId, userName);
 
         public void Dislike(int bookId, string userName)
-        {
-            context.Likes.Remove(GetBookLikeMadeByUser(bookId, userName));
-            context.SaveChanges();
-        }
+            => likeService.Dislike(bookId, userName);
+
+        public void AddComment(int bookId, string userName, string message)
+            => commentService.AddComment(bookId, userName, message);
+
+        public int DeleteCommentById(int id)
+            => commentService.DeleteCommentById(id);
     }
 }
